car-plate-extractor/backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
easyocr==1.7.2
environs==14.1.0
fastapi==0.115.8
fastapi_cors==0.0.6
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.55.8
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
idna==3.10
imageio==2.37.0
imutils==0.5.4
Jinja2==3.1.5
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
ninja==1.11.1.3
numba==0.61.0
numpy==2.0.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-image==0.25.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tifffile==2025.1.10
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.91
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
